Export/import: move file related code to separate package
(no functional change)
