+++ servers/quarkus-cli/src/main/java/org/projectnessie/quarkus/cli/ExportRepository.java
@@ -20,13 +20,13 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Locale;
 import javax.annotation.Nonnull;
-import org.projectnessie.versioned.transfer.ExportFileSupplier;
 import org.projectnessie.versioned.transfer.ExportImportConstants;
-import org.projectnessie.versioned.transfer.FileExporter;
 import org.projectnessie.versioned.transfer.NessieExporter;
 import org.projectnessie.versioned.transfer.ProgressEvent;
 import org.projectnessie.versioned.transfer.ProgressListener;
-import org.projectnessie.versioned.transfer.ZipArchiveExporter;
+import org.projectnessie.versioned.transfer.files.ExportFileSupplier;
+import org.projectnessie.versioned.transfer.files.FileExporter;
+import org.projectnessie.versioned.transfer.files.ZipArchiveExporter;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.ExportMeta;
 import picocli.CommandLine;
 import picocli.CommandLine.PicocliException;
