+++ servers/quarkus-cli/src/main/java/org/projectnessie/quarkus/cli/ImportRepository.java
@@ -34,13 +34,13 @@ import org.projectnessie.versioned.persist.adapter.HeadsAndForkPoints;
 import org.projectnessie.versioned.persist.adapter.ImmutableHeadsAndForkPoints;
 import org.projectnessie.versioned.transfer.CommitLogOptimization;
 import org.projectnessie.versioned.transfer.ExportImportConstants;
-import org.projectnessie.versioned.transfer.FileImporter;
-import org.projectnessie.versioned.transfer.ImportFileSupplier;
 import org.projectnessie.versioned.transfer.ImportResult;
 import org.projectnessie.versioned.transfer.NessieImporter;
 import org.projectnessie.versioned.transfer.ProgressEvent;
 import org.projectnessie.versioned.transfer.ProgressListener;
-import org.projectnessie.versioned.transfer.ZipArchiveImporter;
+import org.projectnessie.versioned.transfer.files.FileImporter;
+import org.projectnessie.versioned.transfer.files.ImportFileSupplier;
+import org.projectnessie.versioned.transfer.files.ZipArchiveImporter;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.ExportMeta;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.HeadsAndForks;
 import picocli.CommandLine;
