+++ versioned/transfer/src/main/java/org/projectnessie/versioned/transfer/ExportCommon.java
@@ -20,6 +20,7 @@ import static org.projectnessie.versioned.transfer.ExportImportConstants.HEADS_A
 
 import java.io.IOException;
 import java.io.OutputStream;
+import org.projectnessie.versioned.transfer.files.ExportFileSupplier;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.ExportMeta;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.ExportVersion;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.HeadsAndForks;
