+++ versioned/transfer/src/main/java/org/projectnessie/versioned/transfer/ExportContext.java
@@ -16,6 +16,7 @@
 package org.projectnessie.versioned.transfer;
 
 import java.io.IOException;
+import org.projectnessie.versioned.transfer.files.ExportFileSupplier;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.Commit;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.ExportMeta;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.NamedReference;
