+++ versioned/transfer/src/main/java/org/projectnessie/versioned/transfer/NessieExporter.java
@@ -23,6 +23,7 @@ import org.immutables.value.Value;
 import org.projectnessie.versioned.StoreWorker;
 import org.projectnessie.versioned.persist.adapter.DatabaseAdapter;
 import org.projectnessie.versioned.store.DefaultStoreWorker;
+import org.projectnessie.versioned.transfer.files.ExportFileSupplier;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.ExportMeta;
 
 @Value.Immutable
