+++ versioned/transfer/src/main/java/org/projectnessie/versioned/transfer/files/ImportFileSupplier.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.versioned.transfer;
+package org.projectnessie.versioned.transfer.files;
 
 import java.io.IOException;
 import java.io.InputStream;
