+++ versioned/transfer/src/test/java/org/projectnessie/versioned/transfer/TestFileExportImport.java
@@ -24,6 +24,10 @@ import java.nio.file.Path;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.projectnessie.versioned.transfer.files.ExportFileSupplier;
+import org.projectnessie.versioned.transfer.files.FileExporter;
+import org.projectnessie.versioned.transfer.files.FileImporter;
+import org.projectnessie.versioned.transfer.files.ImportFileSupplier;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.ExportMeta;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.HeadsAndForks;
