+++ versioned/transfer/src/test/java/org/projectnessie/versioned/transfer/TestZipArchiveExportImport.java
@@ -26,6 +26,10 @@ import java.util.Enumeration;
 import java.util.List;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
+import org.projectnessie.versioned.transfer.files.ExportFileSupplier;
+import org.projectnessie.versioned.transfer.files.ImportFileSupplier;
+import org.projectnessie.versioned.transfer.files.ZipArchiveExporter;
+import org.projectnessie.versioned.transfer.files.ZipArchiveImporter;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.ExportMeta;
 import org.projectnessie.versioned.transfer.serialize.TransferTypes.HeadsAndForks;
